rename project name
Solves: #1
