+++ plugin.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="UTF-8" ?>
-<project name="s3_platform_plugin_template">
+<project name="s3p_plugin_parser_w3c">
     <version>3.0</version>
 </project>
