+++ tests/payload/test_plugin_run.py
@@ -57,7 +57,7 @@ class TestPayloadRun:
 
     def run_payload(self, payload: Type[S3PParserBase], driver: WebDriver, refer: S3PRefer, max_document: int,
                     timeout: int = 2):
-        from src.s3_platform_plugin_template.w3c import W3C
+        from src.s3p_plugin_parser_w3c.w3c import W3C
         if isinstance(payload, type(W3C)):
             _payload = payload(refer=refer, web_driver=driver, max_count_documents=max_document, last_document=None)
 
@@ -72,18 +72,17 @@ class TestPayloadRun:
     def test_run_with_10_docs_restriction(self, chrome_driver, fix_s3pRefer, fix_payload):
         max_docs = 10
         docs = self.run_payload(fix_payload, chrome_driver, fix_s3pRefer, max_docs, 200)
-        assert len(docs) == max_docs
+        assert len(docs) == max_docs, f"Payload вернул {len(docs)} материалов. А должен был {max_docs}"
 
     def test_return_types(self, chrome_driver, fix_s3pRefer, fix_payload):
         max_docs = 10
         docs = self.run_payload(fix_payload, chrome_driver, fix_s3pRefer, max_docs, 200)
-        assert len(docs) == max_docs
-        assert isinstance(docs, tuple) and all([isinstance(el, S3PDocument) for el in docs])
+        assert isinstance(docs, tuple) and all([isinstance(el, S3PDocument) for el in docs]), f""
 
     def test_returned_parameters_are_sufficient(self, chrome_driver, fix_s3pRefer, fix_payload):
         max_docs = 10
         docs = self.run_payload(fix_payload, chrome_driver, fix_s3pRefer, max_docs, 200)
-        assert len(docs) == max_docs
+        assert 0 < len(docs) <= max_docs
         for el in docs:
             assert el.title is not None and isinstance(el.title, str)
             assert el.link is not None and isinstance(el.link, str)
